requirements.txt
pymysql
pillow
flask
opencv
config
authlib